# docs/architecture/openapi/admin-v1.yaml
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Happy Agent Platform Admin API",
    "version": "1.0.0",
    "description": "Closed control-plane contract for typed component catalogs, defaults, Agent lifecycle, evaluation, publication, sessions, runs, traces and playground."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin application"
    }
  ],
  "security": [
    {
      "agentAdminBearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/agents": {
      "get": {
        "tags": [
          "Agents"
        ],
        "operationId": "listAgents",
        "summary": "List Agents",
        "description": "Lists Agent identities, draft versions and current published versions.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "tags": [
          "Agents"
        ],
        "operationId": "createAgent",
        "summary": "Create an Agent and draft",
        "description": "Creates an Agent with the minimum framework, Provider/Model and Prompt selections plus optional Tool and Skill selections.",
        "x-error-codes": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "FRAMEWORK_NOT_SUPPORTED",
          "IDEMPOTENCY_CONFLICT",
          "CONFLICT"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created Agent",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/admin/agents/{agentKey}/draft": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AgentKey"
        }
      ],
      "get": {
        "tags": [
          "Agents"
        ],
        "operationId": "getAgentDraft",
        "summary": "Read the Agent draft",
        "description": "Returns sparse component bindings and explicit overrides; omitted values inherit defaults.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND"
        ],
        "responses": {
          "200": {
            "description": "Agent draft",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDraft"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "patch": {
        "tags": [
          "Agents"
        ],
        "operationId": "updateAgentDraft",
        "summary": "Update the Agent draft",
        "description": "Applies sparse explicit overrides and bindings with optimistic concurrency.",
        "x-optimistic-concurrency": true,
        "x-error-codes": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "FRAMEWORK_NOT_SUPPORTED",
          "PRECONDITION_FAILED",
          "PRECONDITION_REQUIRED",
          "IDEMPOTENCY_CONFLICT"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDraftPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Agent draft",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDraft"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/admin/runs": {
      "get": {
        "tags": [
          "Observability"
        ],
        "operationId": "listRuns",
        "summary": "List Agent runs",
        "description": "Filters durable Runs by Agent, session, status and RFC 3339 time range.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "agentKey",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RunStatus"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/admin/runs/{runId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RunId"
        }
      ],
      "get": {
        "tags": [
          "Observability"
        ],
        "operationId": "getRun",
        "summary": "Read an Agent run",
        "description": "Returns state, budget, token/cost totals, output, error and recovery cursor.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND"
        ],
        "responses": {
          "200": {
            "description": "Run detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/admin/traces/conversations": {
      "get": {
        "tags": ["Observability"],
        "operationId": "adminListRecentConversations",
        "summary": "List recent conversations without exposing user UUID search",
        "description": "Returns the most recently active conversations across users for the personal administrator.",
        "x-error-codes": ["UNAUTHORIZED", "FORBIDDEN"],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 0, "default": 0}},
          {"name": "size", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 30}}
        ],
        "responses": {
          "200": {"description": "Recent conversations", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchConversationSummaryList"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "default": {"$ref": "#/components/responses/Problem"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "agentAdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Platform administrator token only.",
        "x-audience": "happy-agent-admin-v1",
        "x-required-scope": "AGENT_ADMIN"
      }
    },
    "headers": {
      "ETag": {
        "description": "Strong entity tag.",
        "schema": {
          "type": "string",
          "pattern": "^\\\"[A-Za-z0-9._:-]+\\\"$"
        }
      },
      "Location": {
        "description": "Canonical URL of the created resource.",
        "schema": {
          "type": "string",
          "format": "uri-reference"
        }
      },
...
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Caller key scoped to the administrator; a different request digest returns IDEMPOTENCY_CONFLICT.",
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128
        }
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "Strong ETag from the latest resource read.",
        "schema": {
          "type": "string",
          "pattern": "^\\\"[A-Za-z0-9._:-]+\\\"$"
        }
      },
...
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
      "AgentKey": {
        "name": "agentKey",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[a-z][a-z0-9._-]{1,119}$"
        }
      },
...
      "RunId": {
        "name": "runId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "Problem": {
        "description": "Problem Details response.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Malformed request.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Administrator lacks the required scope.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource was not found.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "Resource or idempotency state conflict.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "PreconditionFailed": {
        "description": "If-Match is stale.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unprocessable": {
        "description": "Domain validation failed.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "If-Match is required.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "CursorPageInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "hasMore"
        ],
        "properties": {
          "nextCursor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "Agent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "agentKey",
          "displayName",
          "applicationScope",
          "status",
          "draftRevision",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "agentKey": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "applicationScope": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "DISABLED",
              "RETIRED"
            ]
          },
          "draftRevision": {
            "type": "integer",
            "minimum": 1
          },
          "currentPublishedVersion": {
            "type": "integer",
            "minimum": 1
          },
          "revision": {
            "type": "integer",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AgentPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Agent"
            }
          },
          "page": {
            "$ref": "#/components/schemas/CursorPageInfo"
          }
        }
      },
      "CreateAgentRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "agentKey",
          "displayName",
          "applicationScope",
          "frameworkVersion",
          "providerVersion",
          "model",
          "promptVersion",
          "toolBindings",
          "skillBindings"
        ],
        "properties": {
          "agentKey": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9._-]{1,119}$"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "applicationScope": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "frameworkVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "providerVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "model": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "promptVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "toolBindings": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/ToolBinding"
            }
          },
          "skillBindings": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/SkillBinding"
            }
          }
        }
      },
      "AgentDraft": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "agentKey",
          "revision",
          "frameworkVersion",
          "providerVersion",
          "modelBinding",
          "promptVersion",
          "toolBindings",
          "skillBindings",
          "hookBindings",
          "runtimeOverrides",
          "updatedAt"
        ],
        "properties": {
          "agentKey": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "revision": {
            "type": "integer",
            "minimum": 1
          },
          "frameworkVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "providerVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "modelBinding": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "promptVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "toolBindings": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/ToolBinding"
            }
          },
          "skillBindings": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/SkillBinding"
            }
          },
          "hookBindings": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/HookBinding"
            }
          },
          "memoryPolicyVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "outputSchemaVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "evaluationSuiteVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "defaultProfileVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "runtimeOverrides": {
            "$ref": "#/components/schemas/DefaultValues"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AgentDraftPatchRequest": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "description": "Binding arrays are full replacements. resetPaths removes existing leaf overrides so the next layer is inherited; null is never used.",
        "properties": {
          "frameworkVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "providerVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "modelBinding": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "promptVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "toolBindings": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/ToolBinding"
            }
          },
          "skillBindings": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/SkillBinding"
            }
          },
          "hookBindings": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/HookBinding"
            }
          },
          "memoryPolicyVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "outputSchemaVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "evaluationSuiteVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "defaultProfileVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "runtimeOverrides": {
            "$ref": "#/components/schemas/DefaultValues"
          },
          "resetPaths": {
            "type": "array",
            "uniqueItems": true,
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": [
                "runtime.maxRunSeconds",
                "runtime.maxToolCalls",
                "runtime.maxInputTokens",
                "runtime.maxOutputTokens",
                "runtime.maxCostUsd",
                "runtime.modelParameters.temperature",
                "runtime.modelParameters.topP",
                "runtime.modelParameters.maxOutputTokens",
                "runtime.retryPolicy",
                "memoryPolicyVersion",
                "outputSchemaVersion",
                "evaluationSuiteVersion",
                "defaultProfileVersion"
              ]
            }
          }
        }
      },
...
      "RunStatus": {
        "type": "string",
        "enum": [
          "QUEUED",
          "RUNNING",
          "WAITING_APPROVAL",
          "SUCCEEDED",
          "FAILED",
          "CANCELLED"
        ]
      },
...
      "RunDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "runId",
          "origin",
          "agentKey",
          "status",
          "budget",
          "usage",
          "lastEventId",
          "structuredOutput",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "origin": {
            "type": "string",
            "enum": [
              "SESSION",
              "PLAYGROUND",
              "EVALUATION"
            ]
          },
          "agentKey": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "agentVersion": {
            "type": "integer",
            "minimum": 1
          },
          "draftRevision": {
            "type": "integer",
            "minimum": 1
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "budget": {
            "$ref": "#/components/schemas/RunBudget"
          },
          "usage": {
            "$ref": "#/components/schemas/RunUsage"
          },
          "lastEventId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "structuredOutput": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigEntry"
            }
          },
          "failure": {
            "$ref": "#/components/schemas/JobFailure"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RunPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            }
          },
          "page": {
            "$ref": "#/components/schemas/CursorPageInfo"
          }
        }
      },
...
      "WorkbenchConversationSummaryList": {
        "type": "array",
        "items": {"$ref": "#/components/schemas/WorkbenchConversationSummary"}
      },
...
    }
  }
}